Support 64-bit primitives (#32)
Add support for 64-bit primitives, including fp64, uint64, and int64.
Generally, math operations are performed at emulation precision
then quantized at the end when the value is stored.

Implement support for copyReinterp, which was basically a stub
until now. The function will produce the raw/compressed bits from
the emulated value then cast the bits to the new emulated value,
based on the target type and precision.

It is interesting to note that many GLSL extended set instructions
(especially the transcendental ones) specify that the inputs must be
either 16-bit or 32-bit floats. In those cases, we perform the math
at the lower precision. However, not all instructions (such as Length,
Distance, Cross, etc) have this requirement. For the latter set, we
will now perform the calculation at a higher precision. This matches
what is done for integer instructions, where the result is calculated
exactly then rounded at the end.

Because the GLSL instructions were uniformly being calculated at float32
precision before, we needed to regenerate the output for
procedural-ocean, for which the output was off by a few ULP.

Cooperative matrices follow a slightly different paradigm since we don't
know the number of elements upon construction and they must be dummy
filled separately.

Also, regenerate the output for a test case using the lower precision.

+++ src/values/raytrace/accel-struct.cpp
@@ -114,7 +114,7 @@ void AccelStruct::copyFrom(const Value& new_val) noexcept(false) {
     const Struct& other = Statics::extractStruct(&new_val, "acceleration structure", names);
 
     // tlas: uvec2
-    std::vector<unsigned> tlas_got = Statics::extractUvec(other[0], names[0], 2);
+    std::vector<uint64_t> tlas_got = Statics::extractUvec(other[0], names[0], 2);
     tlas = NodeReference(tlas_got[0], tlas_got[1]);
 
     // Clear any nodes previously held
